requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==4.0.1
attrs==24.1.0
certifi==2024.7.4
exceptiongroup==1.2.2
execnet==2.1.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.2
pytest-order==1.2.1
pytest-xdist==3.6.1
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
